scope all reactabular components from Table instead
