+++ less/patternfly-react.less
@@ -1,4 +1,3 @@
 /**
   Patternfly React Specific Extensions
 */
-@import 'pagination.less';
