--- sass/patternfly-react/_pagination.scss
@@ -1,16 +1,0 @@
-.content-view-pf-pagination {
-  .form-group {
-    .pagination-pf-pagesize.btn-group {
-      display: flex;
-      float: none;
-      margin-left: 0;
-      margin-right: 5px;
-      margin-bottom: 0;
-      padding-right: 13px;
-      width: auto;
-    }
-    .dropdown-menu {
-      min-width: auto;
-    }
-  }
-}
