+++ sass/patternfly-react/_patternfly-react.scss
@@ -1,4 +1,3 @@
 /**
   Patternfly React Partials
 */
-@import '_pagination';
