--- src/components/DataTable/Stories/index.js
@@ -1,9 +1,0 @@
-export {
-  default as clientSortableTableAddWithInfo
-} from './ClientSortableTableStory';
-export {
-  default as clientPaginationTableAddWithInfo
-} from './ClientPaginationTableStory';
-export {
-  default as serverPaginationTableAddWithInfo
-} from './ServerPaginationTableStory.js';
