--- src/components/DataTable/index.js
@@ -1,17 +1,0 @@
-import actionHeaderCellFormatter from './actionHeaderCellFormatter';
-import customHeaderFormattersDefinition from './customHeaderFormattersDefinition';
-import { defaultSortingOrder } from './constants';
-import selectionCellFormatter from './selectionCellFormatter';
-import selectionHeaderCellFormatter from './selectionHeaderCellFormatter';
-import sortableHeaderCellFormatter from './sortableHeaderCellFormatter';
-import tableCellFormatter from './tableCellFormatter';
-
-export {
-  actionHeaderCellFormatter,
-  customHeaderFormattersDefinition,
-  defaultSortingOrder,
-  selectionCellFormatter,
-  selectionHeaderCellFormatter,
-  sortableHeaderCellFormatter,
-  tableCellFormatter
-};
