+++ src/components/Table/Formatters/sortableHeaderCellFormatter.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Table } from '../Table';
+import { Table } from '../index';
 
 const sortableHeaderCellFormatter = ({
   cellProps,
