+++ src/components/Table/Stories/PatternflyTableStory.js
@@ -1,25 +1,41 @@
 import React from 'react';
 import { Table } from '../index';
-import { mockRows } from '../__mocks__/mockRows';
-import { mockPatternflyColumns } from '../__mocks__/mockColumns';
+import { mockBootstrapRows } from '../__mocks__/mockBootstrapRows';
+import { mockPatternflyColumns } from '../__mocks__/mockBootstrapColumns';
+
+import { inlineTemplate } from '../../../../storybook/decorators/storyTemplates';
+import { DOCUMENTATION_URL } from '../../../../storybook/constants';
 
 /**
  * Patternfly Table stories
  */
 
 const patternflyTableAddWithInfo = stories => {
-  stories.addWithInfo('PatternFly Table Styles', '', () => (
-    <div>
-      <h2>
-        PatternFly recommendation: Bootstrap striped, bordered, hover, and
-        responsive
-      </h2>
-      <Table.PfProvider striped bordered hover columns={mockPatternflyColumns}>
-        <Table.Header />
-        <Table.Body rows={mockRows} rowKey="id" />
-      </Table.PfProvider>
-    </div>
-  ));
+  stories.addWithInfo('PatternFly Table Styles', '', () => {
+    let story = (
+      <div>
+        <h2>
+          PatternFly recommendation: Bootstrap striped, bordered, hover, and
+          responsive
+        </h2>
+        <Table.PfProvider
+          striped
+          bordered
+          hover
+          columns={mockPatternflyColumns}
+        >
+          <Table.Header />
+          <Table.Body rows={mockBootstrapRows} rowKey="id" />
+        </Table.PfProvider>
+      </div>
+    );
+    return inlineTemplate({
+      title: 'Table',
+      documentationLink:
+        DOCUMENTATION_URL.PATTERNFLY_ORG_CONTENT_VIEWS + 'table-view/',
+      story: story
+    });
+  });
 };
 
 export default patternflyTableAddWithInfo;
