+++ src/components/Table/Stories/index.js
@@ -1,2 +1,11 @@
 export { default as bootstrapTableAddWithInfo } from './BootstrapTableStory';
 export { default as patternflyTableAddWithInfo } from './PatternflyTableStory';
+export {
+  default as clientSortableTableAddWithInfo
+} from './ClientSortableTableStory';
+export {
+  default as clientPaginationTableAddWithInfo
+} from './ClientPaginationTableStory';
+export {
+  default as serverPaginationTableAddWithInfo
+} from './ServerPaginationTableStory.js';
