+++ src/components/Table/Table.stories.js
@@ -2,20 +2,14 @@ import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
 import {
   bootstrapTableAddWithInfo,
-  patternflyTableAddWithInfo
+  clientSortableTableAddWithInfo,
+  clientPaginationTableAddWithInfo,
+  patternflyTableAddWithInfo,
+  serverPaginationTableAddWithInfo
 } from './Stories';
-import { defaultTemplate } from '../../../storybook/decorators/storyTemplates';
-import { DOCUMENTATION_URL } from '../../../storybook/constants';
 
 const stories = storiesOf('Table', module);
-
 stories.addDecorator(withKnobs);
-stories.addDecorator(
-  defaultTemplate({
-    title: 'Table',
-    documentationLink: DOCUMENTATION_URL.PATTERNFLY_ORG_WIDGETS + '#tables'
-  })
-);
 
 /**
  * Table stories
@@ -23,3 +17,6 @@ stories.addDecorator(
 
 bootstrapTableAddWithInfo(stories);
 patternflyTableAddWithInfo(stories);
+clientSortableTableAddWithInfo(stories);
+clientPaginationTableAddWithInfo(stories);
+serverPaginationTableAddWithInfo(stories);
