+++ src/components/Table/Table.test.js
@@ -2,18 +2,37 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { Table } from './index';
 
-import { mockRows } from './__mocks__/mockRows';
+import { mockBootstrapRows } from './__mocks__/mockBootstrapRows';
 import {
   mockBootstrapColumns,
   mockPatternflyColumns
-} from './__mocks__/mockColumns';
+} from './__mocks__/mockBootstrapColumns';
+
+import { MockClientPaginationTable } from './__mocks__/mockClientPaginationTable';
+import { MockServerPaginationTable } from './__mocks__/mockServerPaginationTable';
+
+test('Mock Client Pagination table renders', () => {
+  const component = renderer.create(
+    <MockClientPaginationTable onRowsLogger={jest.fn()} />
+  );
+  let tree = component.toJSON();
+  expect(tree).toMatchSnapshot();
+});
+
+test('Mock Server Pagination table renders', () => {
+  const component = renderer.create(
+    <MockServerPaginationTable onServerPageLogger={jest.fn()} />
+  );
+  let tree = component.toJSON();
+  expect(tree).toMatchSnapshot();
+});
 
 test('Bootstrap basic table renders properly', () => {
   const component = renderer.create(
     <Table.PfProvider columns={mockBootstrapColumns}>
       <caption>Optional table caption.</caption>
       <Table.Header />
-      <Table.Body rows={mockRows.slice(0, 3)} rowKey="id" />
+      <Table.Body rows={mockBootstrapRows.slice(0, 3)} rowKey="id" />
     </Table.PfProvider>
   );
 
@@ -25,7 +44,7 @@ test('Bootstrap striped table renders properly', () => {
   const component = renderer.create(
     <Table.PfProvider striped columns={mockBootstrapColumns}>
       <Table.Header />
-      <Table.Body rows={mockRows.slice(0, 3)} rowKey="id" />
+      <Table.Body rows={mockBootstrapRows.slice(0, 3)} rowKey="id" />
     </Table.PfProvider>
   );
 
@@ -37,7 +56,7 @@ test('Bootstrap bordered table renders properly', () => {
   const component = renderer.create(
     <Table.PfProvider bordered columns={mockBootstrapColumns}>
       <Table.Header />
-      <Table.Body rows={mockRows.slice(0, 3)} rowKey="id" />
+      <Table.Body rows={mockBootstrapRows.slice(0, 3)} rowKey="id" />
     </Table.PfProvider>
   );
 
@@ -49,7 +68,7 @@ test('Bootstrap hover table renders properly', () => {
   const component = renderer.create(
     <Table.PfProvider hover columns={mockBootstrapColumns}>
       <Table.Header />
-      <Table.Body rows={mockRows.slice(0, 3)} rowKey="id" />
+      <Table.Body rows={mockBootstrapRows.slice(0, 3)} rowKey="id" />
     </Table.PfProvider>
   );
 
@@ -61,7 +80,7 @@ test('Bootstrap condensed table renders properly', () => {
   const component = renderer.create(
     <Table.PfProvider condensed columns={mockBootstrapColumns}>
       <Table.Header />
-      <Table.Body rows={mockRows.slice(0, 3)} rowKey="id" />
+      <Table.Body rows={mockBootstrapRows.slice(0, 3)} rowKey="id" />
     </Table.PfProvider>
   );
 
@@ -74,7 +93,7 @@ test('Bootstrap contextual classes table renders properly', () => {
     <Table.PfProvider condensed columns={mockBootstrapColumns}>
       <Table.Header />
       <Table.Body
-        rows={mockRows}
+        rows={mockBootstrapRows}
         rowKey="id"
         onRow={(row, { rowIndex }) => {
           switch (rowIndex) {
@@ -101,7 +120,7 @@ test('Bootstrap responsive table renders properly', () => {
     <div className="table-responsive">
       <Table.PfProvider columns={mockBootstrapColumns}>
         <Table.Header />
-        <Table.Body rows={mockRows.slice(0, 3)} rowKey="id" />
+        <Table.Body rows={mockBootstrapRows.slice(0, 3)} rowKey="id" />
       </Table.PfProvider>
     </div>
   );
@@ -114,7 +133,7 @@ test('Patternfly table renders properly', () => {
   const component = renderer.create(
     <Table.PfProvider striped bordered hover columns={mockPatternflyColumns}>
       <Table.Header />
-      <Table.Body rows={mockRows} rowKey="id" />
+      <Table.Body rows={mockBootstrapRows} rowKey="id" />
     </Table.PfProvider>
   );
