+++ src/components/Table/__mocks__/mockBootstrapRows.js
@@ -1,0 +1,72 @@
+export const mockBootstrapRows = [
+  {
+    id: 0,
+    first_name: 'Dan',
+    last_name: 'Abramov',
+    username: 'gaearon',
+    commits: 711,
+    additions: 272635,
+    location: 'London, UK',
+    gender: 'male'
+  },
+  {
+    id: 1,
+    first_name: 'Sebastian',
+    last_name: 'Markbåge',
+    username: 'sebmarkbage',
+    commits: 476,
+    additions: 203610,
+    location: 'San Francisco, CA',
+    gender: 'male'
+  },
+  {
+    id: 2,
+    first_name: 'Sophie',
+    last_name: 'Alpert',
+    username: 'sophiebits',
+    commits: 828,
+    additions: 114467,
+    location: 'California',
+    gender: 'female'
+  },
+  {
+    id: 3,
+    first_name: 'Paul',
+    last_name: 'O’Shannessy',
+    username: 'zpao',
+    commits: 820,
+    additions: 87324,
+    location: 'Seattle, WA',
+    gender: 'male'
+  },
+  {
+    id: 4,
+    first_name: 'Pete',
+    last_name: 'Hunt',
+    username: 'petehunt',
+    commits: 205,
+    additions: 86685,
+    location: 'San Francisco, CA',
+    gender: 'male'
+  },
+  {
+    id: 5,
+    first_name: 'Andrew',
+    last_name: 'Clark',
+    username: 'acdlite',
+    commits: 320,
+    additions: 74162,
+    location: 'Redwood City, CA',
+    gender: 'male'
+  },
+  {
+    id: 6,
+    first_name: 'Nathan',
+    last_name: 'Hunzaker',
+    username: 'nhunzaker',
+    commits: 77,
+    additions: 34504,
+    location: 'Durham, NC',
+    gender: 'male'
+  }
+];
