+++ src/components/Table/__mocks__/mockClientPaginationTable.js
@@ -12,15 +12,16 @@ import {
   selectionCellFormatter,
   selectionHeaderCellFormatter,
   sortableHeaderCellFormatter,
-  tableCellFormatter
+  tableCellFormatter,
+  Table,
+  TABLE_SORT_DIRECTION
 } from '../index';
-import { Table, TABLE_SORT_DIRECTION } from '../../Table';
 import { DropdownKebab } from '../../DropdownKebab';
 import { MenuItem } from '../../MenuItem';
 import { Grid } from '../../Grid';
 import { PaginationRow, paginate, PAGINATION_VIEW } from '../../Pagination';
 import { compose } from 'recompose';
-import { mockRows } from './mockDataTableRows';
+import { mockRows } from './mockRows';
 
 /**
  * Reactabular client side paging based on the following api docs:
@@ -454,15 +455,16 @@ import {
   selectionCellFormatter,
   selectionHeaderCellFormatter,
   sortableHeaderCellFormatter,
-  tableCellFormatter
+  tableCellFormatter,
+  Table,
+  TABLE_SORT_DIRECTION
 } from '../index';
-import { Table, TABLE_SORT_DIRECTION } from '../../Table';
 import { DropdownKebab } from '../../DropdownKebab';
 import { MenuItem } from '../../MenuItem';
 import { Grid } from '../../Grid';
 import { PaginationRow, paginate, PAGINATION_VIEW } from '../../Pagination';
 import { compose } from 'recompose';
-import { mockRows } from './mockDataTableRows';
+import { mockRows } from './mockRows';
 
 /**
  * Reactabular client side paging based on the following api docs:
