+++ src/components/Table/__mocks__/mockRows.js
@@ -1,72 +1,225 @@
+/**
+ * mockRows made courtesy of https://swapi.co/
+ */
 export const mockRows = [
   {
     id: 0,
-    first_name: 'Dan',
-    last_name: 'Abramov',
-    username: 'gaearon',
-    commits: 711,
-    additions: 272635,
-    location: 'London, UK',
+    name: 'Luke Skywalker',
+    height: 172,
+    mass: 77,
+    hair_color: 'blond',
+    skin_color: 'fair',
+    eye_color: 'blue',
+    birth_year: '19BBY',
     gender: 'male'
   },
   {
     id: 1,
-    first_name: 'Sebastian',
-    last_name: 'Markbåge',
-    username: 'sebmarkbage',
-    commits: 476,
-    additions: 203610,
-    location: 'San Francisco, CA',
-    gender: 'male'
+    name: 'C-3PO',
+    height: 167,
+    mass: 75,
+    hair_color: 'n/a',
+    skin_color: 'gold',
+    eye_color: 'yellow',
+    birth_year: '112BBY',
+    gender: 'n/a'
   },
   {
     id: 2,
-    first_name: 'Sophie',
-    last_name: 'Alpert',
-    username: 'sophiebits',
-    commits: 828,
-    additions: 114467,
-    location: 'California',
-    gender: 'female'
+    name: 'R2-D2',
+    height: 96,
+    mass: 32,
+    hair_color: 'n/a',
+    skin_color: 'white, blue',
+    eye_color: 'red',
+    birth_year: '33BBY',
+    gender: 'n/a'
   },
   {
     id: 3,
-    first_name: 'Paul',
-    last_name: 'O’Shannessy',
-    username: 'zpao',
-    commits: 820,
-    additions: 87324,
-    location: 'Seattle, WA',
+    name: 'Darth Vader',
+    height: 202,
+    mass: 136,
+    hair_color: 'none',
+    skin_color: 'white',
+    eye_color: 'yellow',
+    birth_year: '41.9BBY',
     gender: 'male'
   },
   {
     id: 4,
-    first_name: 'Pete',
-    last_name: 'Hunt',
-    username: 'petehunt',
-    commits: 205,
-    additions: 86685,
-    location: 'San Francisco, CA',
-    gender: 'male'
+    name: 'Leia Organa',
+    height: 150,
+    mass: 49,
+    hair_color: 'brown',
+    skin_color: 'light',
+    eye_color: 'brown',
+    birth_year: '19BBY',
+    gender: 'female'
   },
   {
     id: 5,
-    first_name: 'Andrew',
-    last_name: 'Clark',
-    username: 'acdlite',
-    commits: 320,
-    additions: 74162,
-    location: 'Redwood City, CA',
+    name: 'Owen Lars',
+    height: 178,
+    mass: 120,
+    hair_color: 'brown, grey',
+    skin_color: 'light',
+    eye_color: 'blue',
+    birth_year: '52BBY',
     gender: 'male'
   },
   {
     id: 6,
-    first_name: 'Nathan',
-    last_name: 'Hunzaker',
-    username: 'nhunzaker',
-    commits: 77,
-    additions: 34504,
-    location: 'Durham, NC',
+    name: 'Beru Whitesun lars',
+    height: 165,
+    mass: 75,
+    hair_color: 'brown',
+    skin_color: 'light',
+    eye_color: 'blue',
+    birth_year: '47BBY',
+    gender: 'female'
+  },
+  {
+    id: 7,
+    name: 'R5-D4',
+    height: 97,
+    mass: 32,
+    hair_color: 'n/a',
+    skin_color: 'white, red',
+    eye_color: 'red',
+    birth_year: 'unknown',
+    gender: 'n/a'
+  },
+  {
+    id: 8,
+    name: 'Biggs Darklighter',
+    height: 183,
+    mass: 84,
+    hair_color: 'black',
+    skin_color: 'light',
+    eye_color: 'brown',
+    birth_year: '24BBY',
+    gender: 'male'
+  },
+  {
+    id: 9,
+    name: 'Obi-Wan Kenobi',
+    height: 182,
+    mass: 77,
+    hair_color: 'auburn, white',
+    skin_color: 'fair',
+    eye_color: 'blue-gray',
+    birth_year: '57BBY',
+    gender: 'male'
+  },
+  {
+    id: 10,
+    name: 'Anakin Skywalker',
+    height: 188,
+    mass: 84,
+    hair_color: 'blond',
+    skin_color: 'fair',
+    eye_color: 'blue',
+    birth_year: '41.9BBY',
+    gender: 'male'
+  },
+  {
+    id: 11,
+    name: 'Wilhuff Tarkin',
+    height: 180,
+    mass: 'unknown',
+    hair_color: 'auburn, grey',
+    skin_color: 'fair',
+    eye_color: 'blue',
+    birth_year: '64BBY',
+    gender: 'male'
+  },
+  {
+    id: 12,
+    name: 'Chewbacca',
+    height: 228,
+    mass: 112,
+    hair_color: 'brown',
+    skin_color: 'unknown',
+    eye_color: 'blue',
+    birth_year: '200BBY',
+    gender: 'male'
+  },
+  {
+    id: 13,
+    name: 'Han Solo',
+    height: 180,
+    mass: 80,
+    hair_color: 'brown',
+    skin_color: 'fair',
+    eye_color: 'brown',
+    birth_year: '29BBY',
+    gender: 'male'
+  },
+  {
+    id: 14,
+    name: 'Greedo',
+    height: 173,
+    mass: 74,
+    hair_color: 'n/a',
+    skin_color: 'green',
+    eye_color: 'black',
+    birth_year: '44BBY',
+    gender: 'male'
+  },
+  {
+    id: 15,
+    name: 'Jabba Desilijic Tiure',
+    height: 175,
+    mass: 1358,
+    hair_color: 'n/a',
+    skin_color: 'green-tan, brown',
+    eye_color: 'orange',
+    birth_year: '600BBY',
+    gender: 'hermaphrodite'
+  },
+  {
+    id: 16,
+    name: 'Wedge Antilles',
+    height: 170,
+    mass: 77,
+    hair_color: 'brown',
+    skin_color: 'fair',
+    eye_color: 'hazel',
+    birth_year: '21BBY',
+    gender: 'male'
+  },
+  {
+    id: 17,
+    name: 'Jek Tono Porkins',
+    height: 180,
+    mass: 110,
+    hair_color: 'brown',
+    skin_color: 'fair',
+    eye_color: 'blue',
+    birth_year: 'unknown',
+    gender: 'male'
+  },
+  {
+    id: 18,
+    name: 'Yoda',
+    height: 66,
+    mass: 17,
+    hair_color: 'white',
+    skin_color: 'green',
+    eye_color: 'brown',
+    birth_year: '896BBY',
+    gender: 'male'
+  },
+  {
+    id: 19,
+    name: 'Palpatine',
+    height: 170,
+    mass: 75,
+    hair_color: 'grey',
+    skin_color: 'pale',
+    eye_color: 'yellow',
+    birth_year: '82BBY',
     gender: 'male'
   }
 ];
