+++ src/components/Table/__mocks__/mockServerPaginationTable.js
@@ -9,9 +9,10 @@ import {
   selectionCellFormatter,
   selectionHeaderCellFormatter,
   sortableHeaderCellFormatter,
-  tableCellFormatter
+  tableCellFormatter,
+  Table,
+  TABLE_SORT_DIRECTION
 } from '../index';
-import { Table, TABLE_SORT_DIRECTION } from '../../Table';
 import { DropdownKebab } from '../../DropdownKebab';
 import { MenuItem } from '../../MenuItem';
 import { Grid } from '../../Grid';
@@ -360,9 +361,10 @@ import {
   selectionCellFormatter,
   selectionHeaderCellFormatter,
   sortableHeaderCellFormatter,
-  tableCellFormatter
+  tableCellFormatter,
+  Table,
+  TABLE_SORT_DIRECTION
 } from '../index';
-import { Table, TABLE_SORT_DIRECTION } from '../../Table';
 import { DropdownKebab } from '../../DropdownKebab';
 import { MenuItem } from '../../MenuItem';
 import { Grid } from '../../Grid';
