+++ src/components/Table/constants.js
@@ -14,3 +14,14 @@ export const TABLE_SORT_DIRECTIONS = [
   TABLE_SORT_DIRECTION.ASC,
   TABLE_SORT_DIRECTION.DESC
 ];
+
+// Reactabular sorting order allows you to specifiy sort asc/desc only and removes
+// the unsorted state. This is consistent with current PF Data Table but should
+// be better spelled out in our design docs.
+// https://github.com/patternfly/patternfly-design/issues/516
+// https://reactabular.js.org/#/data/sorting?a=customizing-sorting-order
+export const defaultSortingOrder = {
+  FIRST: 'asc',
+  asc: 'desc',
+  desc: 'asc'
+};
